+++ tickets-hall/attendance/src/main/java/com/ticketshall/attendance/mq/consumers/TicketConsumer.java
@@ -1,19 +1,26 @@
-package com.ticketshall.attendance.services;
+package com.ticketshall.attendance.mq.consumers;
 
-import com.ticketshall.attendance.events.TicketCreatedEvent;
-import com.ticketshall.attendance.events.TicketDeletedEvent;
-import com.ticketshall.attendance.events.TicketUpdatedEvent;
+import com.ticketshall.attendance.constants.GeneralConstants;
+import com.ticketshall.attendance.entity.id.InboxMessageId;
+import com.ticketshall.attendance.mq.events.TicketCreatedEvent;
+import com.ticketshall.attendance.mq.events.TicketDeletedEvent;
+import com.ticketshall.attendance.mq.events.TicketUpdatedEvent;
 import com.ticketshall.attendance.mapper.TicketMapper;
-import com.ticketshall.attendance.models.Attendee;
-import com.ticketshall.attendance.models.Event;
-import com.ticketshall.attendance.models.Ticket;
+import com.ticketshall.attendance.entity.Attendee;
+import com.ticketshall.attendance.entity.Event;
+import com.ticketshall.attendance.entity.InboxMessage;
+import com.ticketshall.attendance.entity.Ticket;
 import com.ticketshall.attendance.repository.AttendeeRepository;
 import com.ticketshall.attendance.repository.EventRepository;
+import com.ticketshall.attendance.repository.InboxRepository;
 import com.ticketshall.attendance.repository.TicketRepository;
+import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.amqp.rabbit.annotation.RabbitListener;
 import org.springframework.stereotype.Service;
+
+import java.time.LocalDateTime;
 
 @Service
 @RequiredArgsConstructor
@@ -22,10 +29,16 @@ public class TicketConsumer {
     private final AttendeeRepository attendeeRepository;
     private final EventRepository eventRepository;
     private final TicketRepository ticketRepository;
+    private final InboxRepository inboxRepository;
     private final TicketMapper ticketMapper;
 
     @RabbitListener(queues = "${app.rabbitmq.queues.ticketCreated}")
+    @Transactional
     public void handleTicketCreated(TicketCreatedEvent event) {
+        InboxMessageId inboxMessageId = new InboxMessageId(event.id(), GeneralConstants.TICKET_CREATED_INBOX_TYPE);
+        if (inboxRepository.existsById(inboxMessageId)) {
+            return;
+        }
         try {
             log.info("Processing TicketCreatedEvent: {}", event);
 
@@ -40,6 +53,8 @@ public class TicketConsumer {
             ticket.setEvent(ticketEvent);
 
             ticketRepository.save(ticket);
+
+            saveInboxRecord(inboxMessageId);
             log.info("Successfully created ticket with ID: {}", event.id());
         } catch (Exception e) {
             log.error("Error processing TicketCreatedEvent: {}", event, e);
@@ -48,7 +63,12 @@ public class TicketConsumer {
     }
 
     @RabbitListener(queues = "${app.rabbitmq.queues.ticketUpdated}")
+    @Transactional
     public void handleTicketUpdated(TicketUpdatedEvent event) {
+        InboxMessageId inboxMessageId = new InboxMessageId(event.id(), GeneralConstants.TICKET_UPDATED_INBOX_TYPE);
+        if (inboxRepository.existsById(inboxMessageId)) {
+            return;
+        }
         try {
             log.info("Processing TicketUpdatedEvent: {}", event);
             Event ticketEvent = eventRepository.findById(event.eventId())
@@ -62,6 +82,7 @@ public class TicketConsumer {
             ticket.setEvent(ticketEvent);
 
             ticketRepository.save(ticket);
+            saveInboxRecord(inboxMessageId);
         }  catch (Exception e) {
             log.error("Error processing TicketUpdatedEvent: {}", event, e);
             throw e;
@@ -69,12 +90,24 @@ public class TicketConsumer {
     }
     @RabbitListener(queues = "${app.rabbitmq.queues.ticketDeleted}")
     public void handleTicketDeleted(TicketDeletedEvent event) {
+        InboxMessageId inboxMessageId = new InboxMessageId(event.id(), GeneralConstants.TICKET_DELETED_INBOX_TYPE);
+        if (inboxRepository.existsById(inboxMessageId)) {
+            return;
+        }
         try {
             log.info("Processing TicketDeletedEvent: {}", event);
             ticketRepository.deleteById(event.id());
+            saveInboxRecord(inboxMessageId);
         } catch (Exception e) {
             log.error("Error processing TicketDeletedEvent: {}", event, e);
             throw e;
         }
+    }
+    private void saveInboxRecord(InboxMessageId id) {
+        InboxMessage message = InboxMessage.builder()
+                .inboxMessageId(id)
+                .receivedAt(LocalDateTime.now())
+                .build();
+        inboxRepository.save(message);
     }
 }
